+++ src/controllers/BaseController.php
@@ -30,10 +30,10 @@ class BaseController extends Controller
             // Backward-compatibility support for config files, which won't be migrated
             if (!is_array($publisherUserGroup)) {
                 Craft::$app->getDeprecator()->log('publisherUserGroup', 'The `publisherUserGroup` setting has been updated, and will cause a fatal error in Craft 4. Please review our [docs](https://verbb.io/craft-plugins/workflow/docs/get-started/configuration).');
-                
+
                 $publisherUserGroupUid = $publisherUserGroup;
                 $publisherUserGroup = [];
-                
+
                 foreach (Craft::$app->getSites()->getAllSites() as $site) {
                     $publisherUserGroup[$site->uid] = $publisherUserGroupUid;
                 }