+++ src/models/Settings.php
@@ -41,7 +41,7 @@ class Settings extends Model
         $groupUid = $this->editorUserGroup[$site->uid] ?? null;
 
         if ($groupUid) {
-            return Craft::$app->userGroups->getGroupByUid($groupUid);
+            return Craft::$app->getUserGroups()->getGroupByUid($groupUid);
         }
 
         return null;
@@ -65,8 +65,6 @@ class Settings extends Model
             }
 
             $this->reviewerUserGroups = [$site->uid => $uids];
-        } else {
-            $this->reviewerUserGroups = [];
         }
 
         $userGroups = [];
@@ -114,7 +112,7 @@ class Settings extends Model
         $groupUid = $this->publisherUserGroup[$site->uid] ?? null;
 
         if ($groupUid) {
-            return Craft::$app->userGroups->getGroupByUid($groupUid);
+            return Craft::$app->getUserGroups()->getGroupByUid($groupUid);
         }
 
         return null;